Int64 -> Int

+++ test/compiler/codegen.jl
@@ -796,4 +796,4 @@ f48085(@nospecialize x...) = length(x)
 @test Core.Compiler.get_compileable_sig(which(f48085, (Vararg{Any},)), Tuple{typeof(f48085), Int, Vararg{Int}}, Core.svec()) === Tuple{typeof(f48085), Any, Vararg{Any}}
 
 # Make sure that the bounds check is elided in tuple iteration
-@test !occursin("call void @", get_llvm(iterate, Tuple{NTuple{4, Float64}, Int64}))
+@test !occursin("call void @", get_llvm(iterate, Tuple{NTuple{4, Float64}, Int}))
